+++ src/main/java/SG/prueba/Repository/InvoiceDetailsRepository.java
@@ -3,5 +3,5 @@ package SG.prueba.Repository;
 import SG.prueba.DTO.InvoiceDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvoiceDetailsRepository extends JpaRepository<InvoiceDetails, Long> {
+public interface InvoiceDetailsRepository extends JpaRepository<InvoiceDetails, Integer> {
 }
