+++ src/main/java/SG/prueba/Repository/InvoiceRepository.java
@@ -3,5 +3,5 @@ package SG.prueba.Repository;
 import SG.prueba.DTO.Invoices;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvoiceRepository extends JpaRepository<Invoices, Long> {
+public interface InvoiceRepository extends JpaRepository<Invoices, Integer> {
 }
