+++ src/main/java/SG/prueba/Repository/ProductsRepository.java
@@ -3,5 +3,5 @@ package SG.prueba.Repository;
 import SG.prueba.DTO.Products;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductsRepository extends JpaRepository<Products, Long> {
+public interface ProductsRepository extends JpaRepository<Products, Integer> {
 }
